Move gammapy/scripts/cta_sensitivity.py -> gammapy/spectrum/sensitivity.py
