+++ gammapy/scripts/__init__.py
@@ -3,4 +3,3 @@
 Gammapy command line interface (scripts).
 """
 from .spectrum_pipe import *
-from .cta_sensitivity import *
