+++ gammapy/spectrum/__init__.py
@@ -13,3 +13,4 @@ from .extract import *
 from .simulation import *
 from .fit import *
 from .results import *
+from .sensitivity import *
