+++ gammapy/spectrum/sensitivity.py
